+++ src/cqindustry/__init__.py
@@ -2,3 +2,4 @@ from .Base import Base
 from .Platform import Platform
 from .Ring import Ring
 from .Walkway import Walkway
+from .Barrier import jersey_shape, barrier_straight, barrier_cross, barrier_curved, barrier_diagonal, taper_barrier