+++ changes.md
@@ -5,6 +5,7 @@
 * Upped cqterrain dependency version
 * Walkway added irregular grid support
 * Walkway added grid support
+* Added Initial Barrier Design
 
 ## 0.0.1
 * Initial release